tests
Changed handling of mocks and moved the ParameterizedTest to CrawlerTest
